move tests to /tests
